double

+++ Code/src/Player.java
@@ -3,7 +3,7 @@ public class Player {
 	
 	private String name;
 	private int wealth;
-	private int popularity;
+	private double popularity;
 	
 	public int getInfluence(){
 		// TODO add influence calculator
